UKOL_3/scientific-assistant/requirements.txt
langgraph
langchain-anthropic
langchain-core
python-dotenv
wikipedia
